PP-9687 Refactor PaymentRefundsResource
Move code responsible for calling connector to create the refund and
handling the response from connector into a service.
